notification/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
celery==5.4.0
flower==2.0.1
redis==4.6.0
pydantic==2.9.2
pydantic-settings==2.3.3
pyjwt==2.8.0
cryptography==42.0.8
bcrypt==4.2.0
